fix: move workspace preview builder to own file
